feat: 클래스명 수정 및 토큰 검증 예외 처리 구현
