--- src/main/java/umc/catchy/domain/jwt/domain/TokenProvider.java
@@ -1,47 +1,0 @@
-package umc.catchy.domain.jwt.domain;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import java.util.Date;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-import umc.catchy.infra.config.jwt.JWTProperties;
-
-@Component
-@RequiredArgsConstructor
-public class TokenProvider {
-
-    private final JWTProperties jwtProperties;
-
-    public String createToken(String email, Long validity) {
-        Claims claims = Jwts.claims().setSubject(email);
-        Date now = new Date();
-        Date expiryDate = new Date(now.getTime() + validity);
-
-        return Jwts.builder()
-                .setClaims(claims)
-                .setIssuedAt(now)
-                .setExpiration(expiryDate)
-                .signWith(SignatureAlgorithm.HS512, jwtProperties.getSecret())
-                .compact();
-
-    }
-
-    public boolean validateToken(String token) {
-        try {
-            Jwts.parser().setSigningKey(jwtProperties.getSecret()).parseClaimsJws(token);
-            return true;
-        } catch (Exception e) {
-            return false;
-        }
-    }
-
-    public String getEmailFromToken(String token) {
-        Claims claims = Jwts.parser()
-                .setSigningKey(jwtProperties.getSecret())
-                .parseClaimsJws(token)
-                .getBody();
-        return claims.getSubject();
-    }
-}
